+++ fixtures/gists-app/app/routes/index.jsx
@@ -81,6 +81,9 @@ export default function Index() {
               Render error in nested route with ErrorBoundary
             </Link>
           </li>
+          <li>
+            <Link to="/multiple-set-cookies">Multiple Set Cookie Headers</Link>
+          </li>
 
           <li>
             <Link to="prefs">Preferences</Link>