+++ jest.config.js
@@ -6,12 +6,6 @@ module.exports = {
       testMatch: ["<rootDir>/packages/remix-dev/**/*-test.[jt]s?(x)"]
     },
     {
-      displayName: "express",
-      testEnvironment: "node",
-      testMatch: ["<rootDir>/packages/remix-express/**/*-test.[jt]s?(x)"],
-      setupFiles: ["<rootDir>/jest/setupNodeGlobals.ts"]
-    },
-    {
       displayName: "node",
       testEnvironment: "node",
       testMatch: ["<rootDir>/packages/remix-node/**/*-test.[jt]s?(x)"],
@@ -25,6 +19,25 @@ module.exports = {
       ],
       setupFiles: ["<rootDir>/jest/setupNodeGlobals.ts"]
     },
+    // Node Adapters
+    {
+      displayName: "architect",
+      testEnvironment: "node",
+      testMatch: ["<rootDir>/packages/remix-architect/**/*-test.[jt]s?(x)"]
+    },
+    {
+      displayName: "express",
+      testEnvironment: "node",
+      testMatch: ["<rootDir>/packages/remix-express/**/*-test.[jt]s?(x)"],
+      setupFiles: ["<rootDir>/jest/setupNodeGlobals.ts"]
+    },
+    {
+      displayName: "vercel",
+      testEnvironment: "node",
+      testMatch: ["<rootDir>/packages/remix-vercel/**/*-test.[jt]s?(x)"],
+      setupFiles: ["<rootDir>/jest/setupNodeGlobals.ts"]
+    },
+    // Fixture Apps
     {
       displayName: "gists-app",
       testEnvironment: "node",