+++ packages/remix-architect/server.ts
@@ -1,7 +1,14 @@
-import type {
-  Request as ArcRequest,
-  Response as ArcResponse
-} from "@architect/functions";
+import { URL } from "url";
+import {
+  Headers as NodeHeaders,
+  Request as NodeRequest,
+  formatServerError
+} from "@remix-run/node";
+import {
+  APIGatewayProxyEventHeaders,
+  APIGatewayProxyEventV2,
+  APIGatewayProxyHandlerV2
+} from "aws-lambda";
 import type {
   AppLoadContext,
   ServerBuild,
@@ -9,7 +16,6 @@ import type {
 } from "@remix-run/server-runtime";
 import { createRequestHandler as createRemixRequestHandler } from "@remix-run/server-runtime";
 import type { Response as NodeResponse } from "@remix-run/node";
-import { Request as NodeRequest, formatServerError } from "@remix-run/node";
 
 /**
  * A function that returns the value to use as `context` in route `loader` and
@@ -19,7 +25,7 @@ import { Request as NodeRequest, formatServerError } from "@remix-run/node";
  * environment/platform-specific values through to your loader/action.
  */
 export interface GetLoadContextFunction {
-  (req: ArcRequest): AppLoadContext;
+  (event: APIGatewayProxyEventV2): AppLoadContext;
 }
 
 export type RequestHandler = ReturnType<typeof createRequestHandler>;
@@ -36,41 +42,77 @@ export function createRequestHandler({
   build: ServerBuild;
   getLoadContext: GetLoadContextFunction;
   mode?: string;
-}) {
+}): APIGatewayProxyHandlerV2 {
   let platform: ServerPlatform = { formatServerError };
   let handleRequest = createRemixRequestHandler(build, platform, mode);
 
-  return async (req: ArcRequest): Promise<ArcResponse> => {
-    let request = createRemixRequest(req);
+  return async (event, _context) => {
+    let request = createRemixRequest(event);
     let loadContext =
-      typeof getLoadContext === "function" ? getLoadContext(req) : undefined;
+      typeof getLoadContext === "function" ? getLoadContext(event) : undefined;
 
     let response = ((await handleRequest(
       (request as unknown) as Request,
       loadContext
     )) as unknown) as NodeResponse;
 
+    let cookies: string[] = [];
+
+    // Arc/AWS API Gateway will send back set-cookies outside of response headers.
+    for (let [key, values] of Object.entries(response.headers.raw())) {
+      if (key.toLowerCase() === "set-cookie") {
+        for (let value of values) {
+          cookies.push(value);
+        }
+      }
+    }
+
+    if (cookies.length) {
+      response.headers.delete("set-cookie");
+    }
+
     return {
       statusCode: response.status,
       headers: Object.fromEntries(response.headers),
+      cookies,
       body: await response.text()
     };
   };
 }
 
-function createRemixRequest(req: ArcRequest): NodeRequest {
-  let host = req.headers["x-forwarded-host"] || req.headers.host;
-  let search = req.rawQueryString.length ? "?" + req.rawQueryString : "";
-  let url = new URL(req.rawPath + search, `https://${host}`);
+export function createRemixHeaders(
+  requestHeaders: APIGatewayProxyEventHeaders,
+  requestCookies?: string[]
+): NodeHeaders {
+  let headers = new NodeHeaders();
+
+  for (let [header, value] of Object.entries(requestHeaders)) {
+    if (value) {
+      headers.append(header, value);
+    }
+  }
+
+  if (requestCookies) {
+    for (let cookie of requestCookies) {
+      headers.append("Cookie", cookie);
+    }
+  }
+
+  return headers;
+}
+
+export function createRemixRequest(event: APIGatewayProxyEventV2): NodeRequest {
+  let host = event.headers["x-forwarded-host"] || event.headers.host;
+  let proto = event.requestContext.http.protocol || "https";
+  let search = event.rawQueryString.length ? `?${event.rawQueryString}` : "";
+  let url = new URL(event.rawPath + search, `${proto}://${host}`);
 
   return new NodeRequest(url.toString(), {
-    method: req.requestContext.http.method,
-    headers: req.cookies
-      ? { ...req.headers, Cookie: req.cookies.join(";") }
-      : req.headers,
+    method: event.requestContext.http.method,
+    headers: createRemixHeaders(event.headers, event.cookies),
     body:
-      req.body && req.isBase64Encoded
-        ? Buffer.from(req.body, "base64").toString()
-        : req.body
+      event.body && event.isBase64Encoded
+        ? Buffer.from(event.body, "base64").toString()
+        : event.body
   });
 }